+++ src/lib/sparql_base.py
@@ -1,4 +1,5 @@
 from typing import Dict, List
+import requests
 from urllib.error import URLError, HTTPError
 from SPARQLWrapper import SPARQLWrapper, JSON, SPARQLExceptions
 import streamlit as st
@@ -158,7 +159,7 @@ def insert(triples: List[Triple] | Triple, graph: str = None, delete_before=True
     if delete_before and state.get_endpoint().technology == EndpointTechnology.ALLEGROGRAPH:
         delete(triples, graph)
 
-    # If only a single triple is given, transform is into a list
+    # If only a single triple is given, transform it into a list
     if isinstance(triples, Triple):
         triples = [triples]
 
@@ -216,3 +217,26 @@ def delete(triples: List[Triple] | Triple, graph: str = None) -> None:
 
     # Execute
     execute(text)
+
+
+def dump_endpoint() -> str:
+    """Make a full endpoint dump as a n-Quads file."""
+
+    # From state
+    endpoint = state.get_endpoint()
+
+    # Build the URL
+    url = endpoint.url + '/statements'
+
+    # Send GET request with Accept header for N-Quads format
+    headers = {"Accept": "application/n-quads"}
+
+    # Add Authentication
+    if endpoint.username and endpoint.password: auth = (endpoint.username, endpoint.password)
+    elif endpoint.username: auth = (endpoint.username, '')
+    else: auth = None
+
+    # Make the request
+    response = requests.get(url, headers=headers, auth=auth)
+
+    return response.text