+++ src/lib/sparql_shacl.py
@@ -1,11 +1,11 @@
 from typing import List
 import streamlit as st
 from schema import OntologyClass, OntologyProperty, Ontology
+from lib.utils import ensure_uri
 from lib.sparql_base import query
 import lib.state as state
 
 
-@st.cache_data(ttl='1d', show_spinner=False)
 def get_shacl_classes() -> List[OntologyClass]:
     """Get the list of classes listed with the SHACL framework."""
 
@@ -17,7 +17,7 @@ def get_shacl_classes() -> List[OntologyClass]:
             ?uri 
             (COALESCE(?label_, 'Unknown Class Name') as ?label)
         WHERE {
-            """ + ("GRAPH " + endpoint.ontology_uri + " {" if endpoint.ontology_uri else "") + """
+            """ + ("GRAPH " + ensure_uri(endpoint.ontology_uri) + " {" if endpoint.ontology_uri else "") + """
                 ?node a sh:NodeShape .
                 ?node sh:name ?label_ .
                 ?node sh:targetClass ?uri .
@@ -34,7 +34,6 @@ def get_shacl_classes() -> List[OntologyClass]:
     return classes
 
 
-@st.cache_data(ttl='1d', show_spinner=False)
 def get_shacl_properties() -> OntologyProperty:
     """Get the list of properties listed with the SHACL framework."""
     
@@ -43,29 +42,31 @@ def get_shacl_properties() -> OntologyProperty:
 
     text = """
         SELECT
-            ?uri
+            (COALESCE(?target_class_, '') as ?card_of_class_uri)
             (COALESCE(?label_, ?uri) as ?label)
             (COALESCE(?order_, '') as ?order)
             (COALESCE(?min_count_, '') as ?min_count)
             (COALESCE(?max_count_, '') as ?max_count)
-            ?domain_class_uri
-            (COALESCE(?range_class_uri_1, ?range_class_uri_2, '') as ?range_class_uri)
+            (COALESCE(?domain_class_uri_, '') as ?domain_class_uri)
+            ?uri
+            (COALESCE(?range_class_uri_, ?datatype_, '') as ?range_class_uri)
         WHERE {
-            """ + ("GRAPH " + endpoint.ontology_uri + " {" if endpoint.ontology_uri else "") + """
+            """ + ("GRAPH " + ensure_uri(endpoint.ontology_uri) + " {" if endpoint.ontology_uri else "") + """               
                 ?shape sh:property ?node .
-                ?node sh:path ?uri .  
-                ?shape sh:targetClass ?domain_class_uri .
+                ?node sh:path ?supposed_uri .  
+                OPTIONAL { ?shape sh:targetClass ?target_class_ . }
+                OPTIONAL { ?supposed_uri sh:inversePath ?inverse_property_uri . }
                 OPTIONAL { ?node sh:name ?label_ . }
                 OPTIONAL { ?node sh:order ?order_ . }
                 OPTIONAL { ?node sh:minCount ?min_count_ . }
                 OPTIONAL { ?node sh:maxCount ?max_count_ . }
-                OPTIONAL {
-                    ?node sh:class ?range_class_uri_1 .
-                    ?range_shape sh:targetClass ?range_class_uri_1 .
-                }
-                OPTIONAL { ?node sh:datatype ?range_class_uri_2 . }
+                OPTIONAL { ?node sh:datatype ?datatype_ . }
+                OPTIONAL { ?node sh:class ?class . }
+
+                BIND(IF(isBlank(?supposed_uri), '', ?target_class_) as ?domain_class_uri_)
+                BIND(IF(isBlank(?supposed_uri), ?target_class_, ?class) as ?range_class_uri_)
+                BIND(IF(isBlank(?supposed_uri), ?inverse_property_uri, ?supposed_uri) as ?uri)
             """ + ("}" if endpoint.ontology_uri else "") + """
-            FILTER(!isBlank(?uri))
         }
     """